chore: remove custom value library; move Distribution to lib.math
