--- src/main/scala/hydrozoa/lib/cardano/value/package.scala
@@ -1,5 +1,0 @@
-package hydrozoa.lib.cardano
-
-/** Value package. TODO: docs
-  */
-package object value {}
