--- src/test/scala/hydrozoa/lib/cardano/value/Unit.scala
@@ -1,6 +1,0 @@
-package hydrozoa.lib.cardano.value
-
-import org.scalatest.funsuite.AnyFunSuite
-
-/** These tests use the org.scalactic.Equals instances when comparing things with `===`. */
-class Unit extends AnyFunSuite {}
